Fix and change DisplayName
